+++ CHANGES.md
@@ -12,11 +12,12 @@ File offset timestamps, bugfixes, and general maintenance. This version requires
 * Removed features:
   * "Radio paging" was removed (it was only decoded internally anyway). We haven't found any real data
     to test it on and the feature has been deprecated by RDS Forum since 2021.
-* Other fixes:
+* UX and other fixes:
   * Don't print a warning about sample rate when no samples expected (#140).
-  * Print more specific error messages if a file can't be opened (instead of "system error").
+  * Print a warning if expecting raw PCM but WAV header is seen.
+  * Print more specific error messages if a file can't be opened (instead of "system error") or early EOF is encountered.
   * Validate the combination of command-line options more thoroughly, to reduce surprising behavior.
-  * Fix a case where truncated hex input lines were accepted by mistake.
+  * Fix a case where truncated hex input lines were accepted by mistake and zeros generated.
 * Refactoring, CI, etc:
   * Remove support for Ubuntu 20.04 as it's not available any more for CI runners
   * Replaced Debian 10 build check with Debian 11